tmpfs VM, use bridge (static) VM IPs for ssh+pgbench
This required a full copy of the vm_image directory because Dockerfiles
don't allow ADD'ing symlinks that point outside of the context, and
setting the context to the respository's root seemed icky.

The ssh-into-vm.sh and run-bench.sh scripts have been changed so that
they now use the VM's IP address in the `vm-bridge` network.

There was also a weird issue from postgres along the lines of "could not
open shared memory segment" that switching dynamic_shared_memory_type to
mmap fixed.