Add watermark handling
This commit splits the previous notion of "pressure" into two pieces,
"capacity" pressure and "logical" pressure.

Capacity pressure refers to the previous idea of "resources requested
that we couldn't allocate". Logical pressure is the amount of resources
we've reserved beyond our watermark.

Total pressure is equal to capacity presure + logical pressure, so more
migrations occur when the total pressure is greater than the amount
accounted for (which is still equal to reserved + capacity pressure).