agent: Connect to new scheduler on failure
With this change, the autoscaler-agent will not reconnect to a scheduler
that's previously failed, but it *will* connect to a new scheduler once
it comes online.

In general, this seems to work ok. The immediateness of reconnecting is
a litte worrisome, because it *could* amplify existing failures. So...
future work on adding random delays remains.