plugin: Periodically check for deadlocks
Noticed something like this on staging, figured the best course of
action is to actively handle it.

It's theoretically possible to trigger this on super high load, but...
that probably can't happen within the normal constraints of K8s. The
only thing I'd actually be worried about is funky clock stuff messing it
up.