agent: Rewrite runner and informant server (#25)
To signal to git that this is a full rewrite of the related files, they
have different names now. Roughly speaking, the functionality of each
has been moved, following:

* pkg/agent/run.go -> pkg/agent/runner.go
* pkg/agent/serve.go -> pkg/agent/informant.go

The main architectural difference is that the Runner operates as a
singular object with shared state between a collection of worker
threads, where previously some state was shared and other state was sent
over channels.

The new architecture has made it much easier for me to conceptualize
potential issues around racy behavior. Most of the relevant locations
are marked with FIXMEs that I'll open issues (or PRs) for.

There's a *bunch* of other changes here as well. To briefly go through
them:

* Exported some fields of schedulerInfo, so they can be used in the new
  State() method.
* Added a convenience methods to api.Resources
* Added JSON struct tags and convenience methods to api.VmInfo, in
  addition to validation in ExtractVmInfo.
* Updated the autoscaler-agent config to reflect what's actually in use
  (some previously planned future configuration won't be happening)
* Added util.ChanMutex, which provides a sync.Mutex-like interface, but
  with channels (meaning, among other things, it can be select-ed on)

For most of the somewhat-unrelated changes (e.g. ChanMutex), the work of
replacing our usage of equivalent-but-worse constructs (e.g. sync.Mutex)
elsewhere has been left for a follow-up PR.