release: remove `load: true` from docker/build-push-action
We seem to have gotten away with not loading before, and trying to set
both load=true and push=true results in

  Error: buildx failed with: ERROR: push and load may not be set together at the moment

There's basically no information available online about this, and it
*seems* to just be a current limitation of build-push-action? Looking
into the source did not illuminate anything.

We were previously just using push=true (both here and in
neondatabase/neonvm), so this commit switches back to that.