plugin: Don't reject request if VM is constrained (#108)
Ran into this with free tier VMs in production, which are bounded to
exactly 1 CPU and 1 GB memory. The autoscaler-agent (validly) sends a
resource request to the scheduler plugin which wasn't a multiple of the
compute unit (because it can only be 1 CPU + 1 GB).

This commit allows the plugin to recognize that some requests that
aren't a multiple of the compute unit are still valid if it's not
possible for the VM's allocation to be a multiple of the compute unit.

Adds Min and Max fields to podResourceState, to track the information we
need to make the determination.