e2e: run k8s logs/events step if cluster exists
This makes some of the handling around the "Get k8s logs and events"
step a little more precise. We want to run it on failure-ish so long as
the cluster was created.

To do so, this commit makes fetching the logs conditional on cluster
creation *not being skipped* (in case cluster creation failed, or was
cancelled).