agent/billing: panic if VM store unexpectedly stopped (#110)
This should act as a reasonable-ish fail-safe against the VM store
ending early, which would otherwise cause us to continually send the
same consumption metrics (and would be incorrect!).

This is not expected to occur. But it's pretty low-cost to implement,
and we'd like to be very careful about billing data.