make: add targets to build individual docker images (#213)
Just a quality-of-life improvement. When developing locally, I often
only want to rebuild a single image, but changing any file invalidates
the docker caches for all images, so `make docker-build` does a lot of
work that I don't actually care about.

This commit also changes the `docker-build` target so that it no longer
has a dependency on `build`, which *seems* like it's ok.