neonvm: Add current usage annotation to runner Pod
The annotation is based on the value of Spec.Guest.*.Use, not the
status, because that's what the rest of the autoscaling system uses to
make decisions with.

To observe this annotation changing live, you can use:

  kubectl get pod -w -o jsonpath='{.metadata.name}{": "}{.metadata.annotations.vm\\.neon\\.tech/usage}{"\n"}'

(different shells may handle the backslashes differently; the argument
passed to kubectl should have single backslashes, like vm\.neon\.tech)

This feature is probably the easiest path towards patching
cluster-autoscaler to work with our VMs - it avoids needing to pass
through an extra neonvm client everywhere.