agent: Fix deadlock on invalid plugin response
Fixes #219.

What went wrong? When the scheduler plugin returns a response that the
agent thinks is invalid, (*Scheduler).handleFatalError() is called while
holding s.runner.lock, even though handleFatalError() also acquires the
lock.

handleFatalError is sometimes called before the lock has been acquired,
so the easiest way to resolve this is to just release the lock before
calling handleFatalError.

Note: we don't have to worry about different threads racing to call
handleFatalError because the entire operation must take place under the
scope of the request lock - the runner lock is just about being able to
write to some fields in the Scheduler and Runner.