neonvm/api: Add (MilliCPU).AsFloat64() (#261)
In general, resource.Quantity's AsApproximateFloat64() method is less
precise than if we do the conversion ourselves. This commit adds a new
method to do that, and switches to it in all our usage.

---

Originally spotted this in the logs. As a reproduction, take for
example, this program:

    package main

    import (
        "fmt"

        "k8s.io/apimachinery/pkg/api/resource"
    )

    func main() {
        milli := int64(27278)

        q := resource.NewMilliQuantity(milli, resource.BinarySI)
        fmt.Println("resource:", q.AsApproximateFloat64())
        f := float64(milli) / 1000
        fmt.Println("float:", f)
    }

With current K8s versions, this outputs:

    resource: 27.278000000000002
    float: 27.278

(Go playground link: https://go.dev/play/p/S7gW9hxgE7C)