agent: Restart panicked/errored runners (+metrics) (#273)
This doesn't directly fix any individual bug, but it should lessen the
impact of some behaviors we're currently seeing.

The bulk of this commit is adding a new function to `agentState`,
`TriggerRestartIfNecessary`. It is called each time the Runner exits,
and queues restarting the Runner if it failed (from panic or other fatal
error). If the Runner was long-running, it's immediately restarted.

If it has not been up for long, it is restarted after a
random delay of 5-10s.

This commit also:

- Changes `podStatusDump` to have a record of previous "end states"
  (time, error, why it exited)
- Adds a metric `autoscaling_agent_runner_starts`
- Adds a metric `autoscaling_agent_runner_restarts`