deploy: Clean up rbac
This commit retains the original functionality, with some unused objects
removed, some unused permissions revoked, and ClusterRoles standardized.

It also fixes the scheduler not having any permissions relating to
VirtualMachineMigrations, which it will need once those are re-enabled.

This commit removes:

- ClusterRole vm-view
- ClusterRole vm-patcher
- ClusterRoleBinding kube-system/autoscaler-vm-view
- ClusterRoleBinding kube-system/autoscale-scheduler-as-vm-patcher
- Role kube-system/autoscale-scheduler-config-reader
- RoleBinding kube-system/autoscale-scheduler-config-reader

And instead adds:

- NeonVM's various user RBAC ClusterRoles:
  - neonvm-virtualmachine-viewer-role
  - neonvm-virtualmachine-editor-role
  - neonvm-virtualmachinemigration-viewer-role
  - neonvm-virtualmachinemigration-editor-role
- ClusterRoleBinding kube-system/autoscale-scheduler-virtualmachine-viewer
- ClusterRoleBinding kube-system/autoscale-scheduler-virtualmachinemigration-editor
- ClusterRoleBinding kube-system/autoscaler-virtualmachine-editor

The NeonVM RBAC roles were present before, but not included in the
rendered manifests.