neonvm/vm-builder: Fix/improve command passthrough (#263)
Ran into this issue on this PR:
https://github.com/neondatabase/neon/pull/4155

The logs had things like this:

  /neonvm/bin/vmstarter.sh: line 6: /usr/local/bin/compute_ctl-D: not found

This is because the control plane's first argument to compute_ctl is
'-D <database dir>', and it was getting merged with the path to the
binary.

This commit aims to fix that by reworking the current implementation on
the vm-builder side - primarily affecting what vmstart looks like. Now,
.Entrypoint and .Cmd are both double shell-escaped (so that it's only
singly-escaped in vmstarter.sh).

The shell-escaping isn't perfect — in particular, it doesn't handle
backslashes. But it should be good enough for now.

Also changed:

- If no arguments at all are given (via VM spec or original image), then
  vmstart will just exit 1, matching Docker's behavior if you try to
  docker run something with no arguments.
- Whitespace issue should be fixed. echo will naturally put the
  whitespace in the right spots, and >1 space gets collapsed by the
  shell. The important thing is having ≥1 space, not =1 space. There's
  definitely some ways to generate 2 or more spaces, but that's ok.