deploy: Split kustomize for agent and scheduler (#285)
Broadly, this commit does two things:

1. Splits the existing kustomize manifest at ./deploy into one for the
   scheduler and one for the autsocaler-agent (effectively reverts this
   change from #245)
2. Breaks up the individual deploy/*.yaml files so that the differnet
   resources are in different files, like what neonvm has.

The second part is for maintainability. For the first part, there's a
couple reasons:

First, upgrades with breaking changes (like v0.7.0) often require that
the scheduler is upgraded before the autoscaler-agents. We *could* just
accept that the agents wouldn't be working for a bit, but that's... not
ideal, and we can do better.

Second, for testing, it's useful to be able to deploy *just* the
autoscaler-agent or scheduler, particularly when looking at one
component's behavior when the other restarts, or around upgrades.