agent/informant: Do health checks even when suspended (#341)
The health checks are meant to serve as a means of reestablishing
connections if something goes wrong, so that we can have naive "all
errors are fatal" approach in the informant (which allows it to be
simpler - a good thing).

However, in prod I saw a new VM (so: recent informant, that supports
health checks) stop communicating, and narrowed it down to a failure to
"/resume" the autoscaler-agent. When that happened, neither component
sent any more requests because it was waiting for the other. In the
agent's logs, there were repeated messages like:

  Informant health check failed: Informant server is currently suspended

So a reasonable fix for this should be to just be less strict about when
we do health checks.