agent: Add scaling metrics (#334)
Broadly, this commit primarily adds metrics that focus on the change in
resources -- both "requested" changes (to: scheduler, informant, NeonVM)
and "approved" changes (from: scheduler, informant).

The CPU and memory changes are separate metrics. Memory is in units of
Mibibytes. Upscaling vs downscaling is represented by the "direction"
label, which is either "inc" or "dec".

There's also the addition of NeonVM request counts and the associated
result.

In total, the new metrics are:

* autoscaling_agent_scheduler_plugin_{requested,approved}_{cpu,mem}_change_total
* autoscaling_agent_informant_{requested,approved}_{cpu,mem}_change_total
* autoscaling_agent_neonvm_requested_{cpu,mem}_change_total
* autoscaling_agent_neonvm_outbound_requests_total

So, 11 new metrics in total.