plugin: Cleanup state for deleted Nodes
This is a pre-req for adding per-node usage metrics, because of what's
mentioned in #248:

> Looking at the state for prod-us-east-2-delta, there's currently 107
> nodes that the plugin is still tracking state for that don't exist

Unclear whether this will resolve #248 ("scheduler has a memory leak").

---

A lot of the extra work behind this PR is around using the new
watch.Store instead of the preexisting calls to the K8s API when we need
to fetch information about a Node. Realistically... this is probably
more hassle than it's worth, but it's nice to have the consistency.