billing, agent/billing: Log IdempotencyKey of events
In order to make this work on the agent's side, idempotency key
generation needed to be *somewhat* exposed, so this commit adds the
Enrich function, which handles the common "filling out the other fields"
tasks for each event.

By abstracting this interface, we also remove the need for separate
(*Batch).AddAbsoluteEvent vs (*Batch).AddIncrementalEvent methods, so
now we just have a single (*Batch).Add method that just calls Enrich.