informant: Fix parent process stall when child dies qickly
This was... fascinating to debug. Here's a story:

For the past few days in prod, it's seemed like we've had more
"autoscaling stuck" VMs than there should be - even after taking into
account that VMs that are part of a pool will always be stuck.

We finally got confirmation of that with the new metrics from v0.12.0,
which unlocked *just* looking at the "autoscaling stuck" VMs that aren't
in pools, and it turned out there were around 25 — much more than
expected.

So, debugging: the autoscaler-agent's "state dump" feature proves
useful, showing that these are *mostly* VMs from 2023-07-07 and after,
with one from 07-04 and one from 07-05. So maybe a recent release did
something? That wouldn't make sense though — we haven't had any major
changes to the vm-informant, aside from the connection closing fix
(see #367).

Looking at the logs for any of these VMs shows... nothing from the
informant? Very curious. If we go into a VM though, it shows that
there's *one* vm-informant process running though! ... uh oh, that means
it's only *the parent* that's running. And `kill <pid>` doesn't work
(because we're already trapping the signals), but `kill -9 <pid>` does,
and fixes the issue, for one VM.

Now that we know that the parent process is stalled, we can look at the
logs from the *start* of the VM, see when that happened. It turns out we
don't have to look very far — *every single* VM that was affected by
this bug has been affected in exactly the same way. The child process
starts up, dies quickly (because postgres isn't alive yet), and then the
parent process just sits there... waiting.

Thankfully, there's 17 more stuck VMs we can play around with for
debugging! On the next one, let's `kill -6 <pid>` so we can get the
stack traces.

One entry sticks out like a sore thumb:

    goroutine 1 [chan receive, 6640 minutes]:
    runtime.gopark(0x0?, 0x0?, 0x80?, 0x4d?, 0xc00007c2e0?)
        /usr/local/go/src/runtime/proc.go:381 +0xd6 fp=0xc0005777b0 sp=0xc000577790 pc=0x43dab6
    runtime.chanrecv(0xc000089920, 0x0, 0x1)
        /usr/local/go/src/runtime/chan.go:583 +0x49d fp=0xc000577840 sp=0xc0005777b0 pc=0x408c1d
    runtime.chanrecv1(0xc0000ff968?, 0xc0000ff8f8?)
        /usr/local/go/src/runtime/chan.go:442 +0x18 fp=0xc000577868 sp=0xc000577840 pc=0x408718
    main.runRestartOnFailure({0x194a8e0, 0xc0002fef30}, 0x6?, {0xc000340a80, 0x4, 0x4}, {0xc00047ac98, 0x1, 0x1})
        /workspace/cmd/vm-informant/main.go:221 +0x285 fp=0xc000577a20 sp=0xc000577868 pc=0x1429605
    main.main()
        /workspace/cmd/vm-informant/main.go:84 +0x1832 fp=0xc000577f80 sp=0xc000577a20 pc=0x1428cb2
    runtime.main()
        /usr/local/go/src/runtime/proc.go:250 +0x207 fp=0xc000577fe0 sp=0xc000577f80 pc=0x43d687
    runtime.goexit()
        /usr/local/go/src/runtime/asm_amd64.s:1598 +0x1 fp=0xc000577fe8 sp=0xc000577fe0 pc=0x471d21

Specifically, it was stuck trying to receive on a channel — the one for
the timer.

It turns out that the typical advice of:

    if !timer.Stop() {
        <-timer.C
    }

.. doesn't apply if you've *already received from the channel*. Because
in that case, having timer.Stop() return false means that the timer
already finished, which would make sense (you already received on it),
and so it isn't going to put another item into the channel for you.

In our case, we received from the channel on a previous iteration of the
loop, and then try to receive from it again when the child informant
dies for a *second* time.

So the fix: Track if you've received from the channel; don't block on
receiving from timer.C if there won't be anything there.