plugin: Fix behavior of ignored namespaces during Filter (#416)
The intention behind that piece of #399 was that pods from ignored
namespaces (particularly: 'overprovisioning') should be included *only*
in the Filter calculations, so that it gets used to trigger eviction for
those pods, but NOT migrations.

However, the current behavior (before this PR) just ignores all pods
that weren't previously known, so the ignored namespaces were *still*
getting ignored during filter. That's fixed here (missed it in #399).