neonvm: Add user-defined additional InitContainers
To do this, we add a new field to the VM spec, ExtraInitContainers.

In order to make this fully backwards- and forwards-compatible with
existing objects (so we don't end up performing a bunch of writes
immediately after deploying), the field is set to "omitempty" so that
existing VMs will be unchanged.