plugin: Handle VM 'Buffer' on delete
Noticed this because 'Buffer' was high in prod-eu-central-1-gamma after
the scheduler restarted, realized it was because a VM was deleted before
the autoscaler-agent checked in about it, and the VM's Buffer wasn't
appropriately removed from the Node's total.

AFAICT this wasn't actually affecting any scheduling decisions (because
Buffer is included in Reserved, and Reserved *is* appropriately unset
when a VM is deleted), but it's worth fixing anyways so that our
monitoring is cleaner.