agent: Improve "billed stuck" metric (#425)
With the way that the informant currently works, when pooled VMs have
autoscaling enabled they're marked as "autoscaling stuck". Because of
that, there's the occasional VM that is detected as "autoscaling stuck"
*after* it's assigned an endpoint, but before the agent has tried to
reconnect to its informant.

If we track the "billed autoscaling stuck" metric with a grace period
extending from the time at which the VM was assigned the endpoint, we
should have enough wiggle room to make the metrics more consistent.