util/watch, agent/billing: Don't record events if VM watch is failing (#439)
On the 'util/watch' side: This adds a new Failing() method that
complements the existing Stopped() method, similarly matching the
metrics we expose.

On the 'agent/billing' side: If the watch is failing, we just pretend
that there's no VMs, and log a suitable error.

Realized this needed to be implemented because we had a k8s API server
switchover in prod-us-west-2-eta, and all the Watch instances separately
failed for 1s at different times. If it had been longer (or if relisting
took a long time), and we had VMs there, we would have used stale data
for our billing events.