Use a fixed port for the informant -> agent connections. (#447)
* Use a fixed port for the informant -> agent connections.

Previously, the agent assigned a new random agent ID, and opened a new
random port dedicated each informant. It then calls the informant's
/register endpoint, giving it the agent ID and the port. From there
on, the informant makes calls (HTTP requests) to that port whenever it
wants to request upscaling, and periodically to check that the
registration is still valid.

Having a dedicated port for each informant causes some problems:

1. It puts a limit on how many informants the agent can manage
concurrently. Each informant requires a dedicated TCP port, and
there's only so many of them.

2. It makes it hard to restrict network connections from the
informant, when it might need to connect to an arbitrary port.

To alleviate those problems, listen for informant -> agent requests on
one, fixed port (10302). Previously, in the /register request, the
agent gave the informant an address like "<ip address>:<port>". We can
multiplex the requests to a single port by including an URL path
component in the address. The agentServeAddr now looks like "<ip
address>:10302/<random mux ID>". So instead of allocating a dedicated
port for each informant, we allocate just a random string (mux ID)
instead.

This works without any changes to the informant, because the informant
doesn't parse agentServeAddr. It simply forms the final url like
"http://<agentServeAddr>/<path>", so agentServeAddr can contain an
additional path prefix and you still get a valid URL.

This commit introduces a fairly generic HttpMuxServer, which routes
HTTP requests to multiple http.ServeMux instances based on the path
prefix. This minimizes the code changes to the rest of the code.
InformantServer creates a new http.ServeMux object, with paths like
"/id", "/resume" etc. just like before, but instead of creating a new
HTTP server, it registers the http.ServeMux with the global
HttpMuxServer, with a random prefix. And instead of shutting down the
dedicated HTTP listener, it unregisters itself from the global
HttpMuxServer.

Fixes https://github.com/neondatabase/autoscaling/issues/420