neonvm: Make spec.EnableAcceleration a pointer
AFAICT, I ran into an issue with neondatabase/cloud#6313 because v0.8.0
added the EnableAcceleration field which *technically* defaults to true,
but without any omitempty the "default" spec sets the field to false and
explicitly provides that to Kubernetes, overriding the intended default.

This is a similar change to #257, for the most part.