neonvm/controller: Deduplicate QMP query logic (#461)
The general idea here is to make all of the functions take only the
information they need - so if they only need the address to connect to,
don't give them the whole VirtualMachine object. This allows us to
remove some functions that had duplicated logic depending on whether
there was a VM as input or the IP + port.

To compensate and provide *some* remaining ease of use, there's a new
QmpAddr function that gets the address of a VM's QMP listener. Because
of how go handles returning multiple values (and passing that along into
another function), we can just directly call functions with the output
of QmpAddr, like:

  QmpQueryMemoryDevices(QmpAddr(vm))

even though QmpAddr returns (string, int32) as the (IP, port) combo.