propagate informant server contexts to dispatcher (#494)
This should ensure that if the outer context is cancelled (for
example, the informant server shutting down), we (the goroutine doing
`disp.run(ctx)`) also shut down. 

Not having this previously caused an error where
the informant server would shut down, close the websocket, and then
because the `disp.run` goroutine was not hooked into the context, it
wouldn't be shut down and would try to read from the closed connection.