runner: separate qmp socket for the SIGTERM handler goroutine
Before this PR, an existing QMP connection, e.g., from neonvm-controller
would cause dial timeout for the runner's SIGTERM handler.

The reason is that a single `-qmp` instance handles connections
sequentially (trivial to test this with one qemu proces and two other
terminals, each connecting to the qmp socket).

The solution in this PR is to add a second `-qmp` instance, just for
the SIGTERM handler, that listens on a UNIX domain socket.

Related PRs:
- neonvm-controller fix to avoid leaking connections:
  https://github.com/neondatabase/autoscaling/pull/527
- actually correct handling of the ACPI power button event
  that our SIGTERM handler would trigger, were it able to
  connect to the QMP socket
  https://github.com/neondatabase/autoscaling/pull/525/