respect vm-monitor's denied downscaling
This commit was actually really fun to write. By doing it this way, we
actually automatically get "retry with a slightly larger downscale" for
free, and hooking everything up was super simple!

Also, while we're here, it's worth adding back the other tests from
before this PR, now that they pass.