Fix some exhaustruct lints
Noticed while working on #371, but it appears that golangci-lint
will sometimes hide errors that it *should* be showing if compiling
succeeds. So if, hypothetically, you're impacted by golang/go#58218 and
normally you're able to get by with 'go build -buildvcs=false ./...' but
golangci-lint doesn't have that functionality, then this is able to
reproduce on main.