switch from IDs to generation numbers (+ schedulerGone)
With IDs, it was theoretically possible for us to reconnect to the same
scheduler instance after disconnecting, which would have the same IDs.
We could have mitigated this by including the scheduler's
resourceVersion in the ID, but IDs are a little hard to grok anyways and
tend to require spooky action at a distance, so generation numbers it is!

---

also forwards scheduler deletion in trackSchedulerLoop into
SchedulerGone calls in the executor.