agent: Fix oversized websocket close frames
As written in the added comment, there's a size limit for our websocket
close frames, and we often exceed that on these "Error handling message"
errors. When that happens, the weboscket library emits a non-JSON log
that we have no way of specifying an alternate sink for, unfortunately.

So, to resolve all of the above, this commit changes the "reason" for
the close frame on those errors to just have the "root" error, rather
than the entire thing.