plugin: Fix migration source deletion event kind
In short: after #563, deleting a migration before completion has caused
the scheduler plugin to incorrectly interpret the deletion event for the
source (or target? not entirely sure) pod as referring to a pod (rather
than a VM), because all pods related to a VM will have the
vm.neon.tech/name label, but only the pod currently _running_ the VM
will be marked as "owned" by the VM; the other will be owned by the
migration.