agent: Expose per-VM size metrics on :9101
Adds a single new metric `autoscaling_vm_resources`, exposed on a
different port because it's expected to be high cardinality.

The metric is parameterized by the resource ("cpu" / "mem") and by what
the value represents ("min" / "spec_use" / "status_use" / "max").

CPU values are in units of CPUs, memory values are in units of bytes.