neonvm: Allow per-VM custom kernel images
As part of debugging neondatabase/neon#5678, we want to test certain
kernel configurations. Currently, this requires rebuilding the kernel
and redploying neonvm-controller with an updated neonvm-runner that
embeds the new kernel. That's too much effort.

So instead, we'd like to be able to use a custom kernel on a per-VM
basis. Because this *generally* should not be the common path, this PR
adds that as a new init container to the runner pod, using the same
"download via image" that we do for root disks.