vm-builder: Use same output format as `docker build` (#605)
We've periodically opened PRs to slightly improve the output of
vm-builder by handling more message types (#280, #599). But the actual
implementation that `docker build` uses is available for us, so we might
as well use that instead.

In the future, it'd be cool to use buildkit here to mimic the better
styling from `docker buildx build`, but its source code is much harder
to navigate.