vm-builder: Set vector scrape interval from 15s to 4s
Per the comment: The default interval is every 15s, but we scrape every
5s in the autoscaler-agent.

This should have the net effect of making the scaling algorithm more
responsive, because we should stop getting (as many) sequential
responses that are exactly the same.

Unfortunately, load average is only recalculated every 5 seconds
anyways, so we may still get repeated values there, but this should at
least be an improvement.