neonvm/controller: Don't delete runner pods in CI
This is implemented by adding a build tag 'nodelete', which disables
deletion of VM runner pods - at least, while the VM or VMM object
exists.

We *could* have actually kept the runner pods around even if the VM/VMM
is deleted, there's already ways to keep the objects around, and this
way the results would be a little less catastrophic if we accidentally
released this.

The 'nodelete' tag is enabled by the Makefile iff CI.