agent/schedwatch: Replace trackcurrent with global value (#675)
ref https://github.com/neondatabase/autoscaling/issues/665#issuecomment-1842146035

In short: We've hit *many* issues with the existing `trackcurrent`
setup, so now that we've merged some changes to improve the semantics of
agent->scheduler requests, we can completely replace this complicated
state machine with a global value.

This commit also removes the scheduler lifecycle tracking from
pkg/agent/{execbridge,executor,core} - which is a big simplification as
well.