CI: run e2e tests only using k3d cluster (#678)
Since we started to run e2e tests using both `kind` and `k3d` [0], I can't recall a
case when tests failed only on one cluster (due to app changes and not a
cluster configuration).

After switching to `large` self-hosted runners (in [1]), there's a chance
that e2e tests could fight for resources with other projects that use
such runners. To make the build faster by using fewer runners. Given the
message above, let's run the test only on a single cluster configuration
but keep an alternative way (for those who are used to it).

It is still possible to trigger e2e tests on `kind` cluster manually.

- [0] https://github.com/neondatabase/autoscaling/pull/340
- [1] https://github.com/neondatabase/autoscaling/pull/661