agent/billing: Log URL on billing requests
ref https://neondb.slack.com/archives/C03H1K0PGKH/p1702049320084879?thread_ts=1701881238.622479

To do this accurately, this commit renames `billing.Client.{BaseURL => URL}`
and adds the `usage_events` suffix to it.

This commit also reorders some of the log fields in the sending code
(mostly just moving traceID), so that smaller items are at the start and
similar items are grouped together.