agent,plugin: Don't send ComputeUnit from plugin
Instead, define it in the autoscaler-agent config.

This change starts the removal process of ComputeUnit from
PluginResponse, so that the new protocol version expects that it won't
be set - but to preserve backwards-compatibility during the upgrade, we
need to keep the old behavior for now (hence the field is `omitempty`).

A follow-up change will update the plugin so that it *never* sends
ComputeUnit, removing the field and the plugin's config along with it.