agent/core: Send monitor requests in 1 CU increments
Before, we would request the "full" downscaling or upscaling that we
desired - for downscaling, we'd retry with a higher amount if the
initial request was denied.

To reduce load, it's better to only request decreasing by 1 CU at a
time. Also, bounding the size of changes from vm-monitor requests will
also help with managing uncertainty.