util/watch: Add relist rate limiting
For the scheduler, we're mostly fine without this (in practice, rate of
relists is quite low).
This *will* be required for the autoscaler-agent as we try to use the VM
spec as a source of truth; we need relist rate limiting because we need
relists in order to make a consistent read of the source of truth.