neonvm: Use crictl to change container CPU, ditch cgroup (#738)
NB: This PR is conditionally enabled via the --enable-container-mgr flag
on neonvm-controller. There are no effects without that.

---

We recently realized[^1] that under cgroups v2, kubernetes uses cgroup
namespaces which has a few effects:

1. The output of /proc/self/cgroup shows as if the container were at the
   root of the hierarchy
2. It's very difficult for us to determine the actual cgroup that the
   container corresponds to on the host
3. We still can't directly create a cgroup in the container's namespace
   because /sys/fs/cgroup is mounted read-only

So, neonvm-runner currently *does not* work as expected with cgroups v2;
it creates a new cgroup for the VM, at the top of the hierarchy, and
doesn't clean it up on exit.

How do we fix this? The aim of this PR is to remove the special cgroup
handling entirely, and "just" go through the Container Runtime Interface
(CRI) exposed by containerd to modify the existing container we're
running in.

This requires access to /run/containerd/containerd.sock, which a
malicious user could use to perform priviledged operations on the host
(or in any other container on the host).
Obviously we'd like to prevent that as much as possible, so the CPU
handling is now runs alongside neonvm-runner as a separate container.
neonvm-runner does not have access to the containerd socket.

On the upside, one key benefit we get from this is being able to set cpu
shares, the abstraction underlying container resources.requests.
The other options weren't looking so great[^2], so if this works, this
would be a nice compromise.

[^1]: https://neondb.slack.com/archives/C03TN5G758R/p1705092611188719
[^2]: https://github.com/neondatabase/autoscaling/issues/591